Move NavigationBar into /components and update outdated imports
